add: permission middleware add

+++ routes/api.php
@@ -2,7 +2,7 @@
 
 use App\Http\Controllers\Api\AdminUserController;
 use App\Http\Controllers\Api\AuthController;
-use App\Http\Controllers\Role\RoleController;
+use App\Http\Controllers\Api\Role\RoleController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
@@ -33,18 +33,18 @@ Route::prefix('v1')->group(function () {
 
     Route::middleware('auth:sanctum')->group(function () {
         Route::controller(AdminUserController::class)->prefix("/admin-user")->group(function () {
-            Route::get('/', 'userList');
-            Route::get('/detail/{adminUserId}', 'userDetail');
-            Route::put('/update/{adminUser}', 'userUpdate');
-            Route::delete('/delete/{adminUser}', 'userDelete');
+            Route::get('/', 'userList')->middleware('permission:user,view');
+            Route::get('/detail/{adminUserId}', 'userDetail')->middleware('permission:user,view');
+            Route::put('/update/{adminUser}', 'userUpdate')->middleware('permission:user,update');
+            Route::delete('/delete/{adminUser}', 'userDelete')->middleware('permission:user,delete');
         });
 
         Route::controller(RoleController::class)->prefix('/role')->group(function () {
-            Route::get('/permissions', 'roleAndPermissionsList');
-            Route::get('/', 'roleList');
-            Route::post('/store', 'roleStore');
-            Route::put('/update/{role}', 'roleUpdate');
-            Route::delete('/delete/{role}', 'roleDelete');
+            Route::get('/permissions', 'roleAndPermissionsList')->middleware('permission:role,view');
+            Route::get('/', 'roleList')->middleware('permission:role,view');
+            Route::post('/store', 'roleStore')->middleware('permission:role,create');
+            Route::put('/update/{role}', 'roleUpdate')->middleware('permission:role,update');
+            Route::delete('/delete/{role}', 'roleDelete')->middleware('permission:role,delete');
         });
     });
